requirements.txt
aiogram>=3.0.0
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
asyncpg
pydantic-settings
bcrypt
pyjwt
python-multipart
aiofiles
jinja2
